tvshows

+++ app/Http/Controllers/AdminSeriesController.php
@@ -26,6 +26,7 @@ use Auth;
 use Hash;
 use Illuminate\Support\Facades\Cache;
 use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\File;
 use View;
 use Validator;
 use ProtoneMedia\LaravelFFMpeg\Support\FFMpeg as FFMpeg;
@@ -43,7 +44,6 @@ use App\Artist;
 use App\Seriesartist;
 use GifCreator\GifCreator;
 use FFMpeg\Coordinate\TimeCode;
-use File;
 use App\SeriesCategory as SeriesCategory;
 use App\SeriesLanguage as SeriesLanguage;
 use GuzzleHttp\Client;
@@ -2297,7 +2297,9 @@ class AdminSeriesController extends Controller
             }
 
             foreach ($files as $file) {
-                unlink($file);
+                if (File::exists($file)) {
+                    unlink($file);
+                } 
             }
 
             Episode::destroy($id);
